requirements.txt
lib-version @ git+https://github.com/remla-rstular/lib-version@v0.0.3
fastapi==0.115.12
Jinja2==3.1.6
uvicorn==0.34.2
aiodns==3.3.0
aiohttp==3.11.18
pydantic-settings==2.9.1
sqlmodel==0.0.24
